requirements.txt
pygame-ce
pygame-gui
opencv-python
numpy